rh: moving from int to uint32 indices

+++ protocols/randhound/utils.go
@@ -9,7 +9,7 @@ import (
 	"github.com/dedis/crypto/random"
 )
 
-func (rh *RandHound) chooseTrustees(Rc, Rs []byte) (map[int]int, []abstract.Point) {
+func (rh *RandHound) chooseTrustees(Rc, Rs []byte) (map[uint32]uint32, []abstract.Point) {
 
 	// Seed PRNG for selection of trustees
 	var seed []byte
@@ -18,12 +18,12 @@ func (rh *RandHound) chooseTrustees(Rc, Rs []byte) (map[int]int, []abstract.Poin
 	prng := rh.Node.Suite().Cipher(seed)
 
 	// Choose trustees uniquely
-	shareIdx := make(map[int]int)
+	shareIdx := make(map[uint32]uint32)
 	trustees := make([]abstract.Point, rh.Group.K)
 	tns := rh.Tree().ListNodes()
-	j := 0
-	for len(shareIdx) < rh.Group.K {
-		i := int(random.Uint64(prng) % uint64(len(tns)))
+	j := uint32(0)
+	for uint32(len(shareIdx)) < rh.Group.K {
+		i := uint32(random.Uint64(prng) % uint64(len(tns)))
 		// Add trustee only if not done so before; choosing yourself as an trustee is fine; ignore leader at index 0
 		if _, ok := shareIdx[i]; !ok && !tns[i].IsRoot() {
 			shareIdx[i] = j // j is the share index
@@ -40,13 +40,13 @@ func (rh *RandHound) hash(bytes ...[]byte) []byte {
 
 func (rh *RandHound) newGroup(nodes int, trustees int) (*Group, []byte, error) {
 
-	n := nodes    // Number of nodes (peers + leader)
-	k := trustees // Number of trustees (= shares generaetd per peer)
+	n := uint32(nodes)    // Number of nodes (peers + leader)
+	k := uint32(trustees) // Number of trustees (= shares generaetd per peer)
 	buf := new(bytes.Buffer)
 
 	// Setup group parameters: note that T <= R <= K must hold;
 	// T = R for simplicity, might change later
-	gp := [6]int{
+	gp := [6]uint32{
 		n,           // N: total number of nodes (peers + leader)
 		n / 3,       // F: maximum number of Byzantine nodes tolerated
 		n - (n / 3), // L: minimum number of non-Byzantine nodes required
@@ -68,7 +68,7 @@ func (rh *RandHound) newGroup(nodes int, trustees int) (*Group, []byte, error) {
 
 	// Include group parameters into group ID
 	for _, g := range gp {
-		if err := binary.Write(buf, binary.LittleEndian, uint32(g)); err != nil {
+		if err := binary.Write(buf, binary.LittleEndian, g); err != nil {
 			return nil, nil, err
 		}
 	}
@@ -100,8 +100,8 @@ func (rh *RandHound) newSession(public abstract.Point, purpose string, time time
 		Time:        time}, rh.hash(pub, []byte(purpose), tm), nil
 }
 
-func (rh *RandHound) nodeIdx() int {
-	return rh.Node.TreeNode().EntityIdx
+func (rh *RandHound) nodeIdx() uint32 {
+	return uint32(rh.Node.TreeNode().EntityIdx)
 }
 
 func (rh *RandHound) sendToChildren(msg interface{}) error {
